refactor: move LOCAL_NETWORK_NAME
