+++ src/ape/utils/misc.py
@@ -40,6 +40,7 @@ DEFAULT_LOCAL_TRANSACTION_ACCEPTANCE_TIMEOUT = 20
 DEFAULT_LIVE_NETWORK_BASE_FEE_MULTIPLIER = 1.4
 DEFAULT_TRANSACTION_TYPE = 0
 DEFAULT_MAX_RETRIES_TX = 20
+LOCAL_NETWORK_NAME = "local"
 SOURCE_EXCLUDE_PATTERNS = (
     ".build",
     ".cache",
