+++ src/ape_ethereum/ecosystem.py
@@ -25,8 +25,11 @@ from ethpm_types.abi import ABIType, ConstructorABI, EventABI, MethodABI
 from pydantic import Field, computed_field, field_validator, model_validator
 from pydantic_settings import SettingsConfigDict
 
-from ape.api import BlockAPI, EcosystemAPI, PluginConfig, ReceiptAPI, TraceAPI, TransactionAPI
-from ape.api.networks import LOCAL_NETWORK_NAME
+from ape.api.config import PluginConfig
+from ape.api.networks import EcosystemAPI
+from ape.api.providers import BlockAPI
+from ape.api.trace import TraceAPI
+from ape.api.transactions import ReceiptAPI, TransactionAPI
 from ape.contracts.base import ContractCall
 from ape.exceptions import (
     ApeException,
@@ -49,20 +52,18 @@ from ape.types import (
     RawAddress,
     TransactionSignature,
 )
-from ape.utils import (
+from ape.utils.abi import LogInputABICollection, Struct, StructParser, is_array, returns_array
+from ape.utils.basemodel import _assert_not_ipython_check, only_raise_attribute_error
+from ape.utils.misc import (
     DEFAULT_LIVE_NETWORK_BASE_FEE_MULTIPLIER,
     DEFAULT_LOCAL_TRANSACTION_ACCEPTANCE_TIMEOUT,
+    DEFAULT_MAX_RETRIES_TX,
     DEFAULT_TRANSACTION_ACCEPTANCE_TIMEOUT,
+    DEFAULT_TRANSACTION_TYPE,
     EMPTY_BYTES32,
+    LOCAL_NETWORK_NAME,
     ZERO_ADDRESS,
-    LogInputABICollection,
-    Struct,
-    StructParser,
-    is_array,
-    returns_array,
 )
-from ape.utils.basemodel import _assert_not_ipython_check, only_raise_attribute_error
-from ape.utils.misc import DEFAULT_MAX_RETRIES_TX, DEFAULT_TRANSACTION_TYPE
 from ape_ethereum.proxies import (
     IMPLEMENTATION_ABI,
     MASTER_COPY_ABI,
